+++ src/main/java/kitchenpos/menu/ui/request/MenuProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.request;
+package kitchenpos.menu.ui.request;
 
 import kitchenpos.menu.domain.MenuProduct;
