+++ src/test/java/kitchenpos/integration/MenuIntegrationTest.java
@@ -8,7 +8,7 @@ import kitchenpos.integration.api.MenuApi;
 import kitchenpos.integration.api.MenuGroupApi;
 import kitchenpos.integration.api.ProductApi;
 import kitchenpos.integration.utils.MockMvcResponse;
-import kitchenpos.ui.request.MenuProductRequest;
+import kitchenpos.menu.ui.request.MenuProductRequest;
 import kitchenpos.menu.application.response.MenuResponse;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
