+++ src/test/java/kitchenpos/integration/api/MenuApi.java
@@ -5,8 +5,8 @@ import java.util.List;
 import kitchenpos.integration.api.texture.ProductTexture;
 import kitchenpos.integration.utils.MockMvcResponse;
 import kitchenpos.integration.utils.MockMvcUtils;
-import kitchenpos.ui.request.MenuCreateRequest;
-import kitchenpos.ui.request.MenuProductRequest;
+import kitchenpos.menu.ui.request.MenuCreateRequest;
+import kitchenpos.menu.ui.request.MenuProductRequest;
 import kitchenpos.menu.application.response.MenuResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
