+++ src/test/java/kitchenpos/integration/api/MenuGroupApi.java
@@ -3,7 +3,7 @@ package kitchenpos.integration.api;
 import java.util.List;
 import kitchenpos.integration.utils.MockMvcResponse;
 import kitchenpos.integration.utils.MockMvcUtils;
-import kitchenpos.ui.request.MenuGroupCreateRequest;
+import kitchenpos.menu.ui.request.MenuGroupCreateRequest;
 import kitchenpos.menu.application.response.MenuGroupResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
